Fixed curl syntax. Checking for linux only right now. Added comments.

+++ apps/appd/appd-region-post-init.sh
@@ -4,18 +4,29 @@
 
 env
 
-yum install -y openssh-clients
+if [ "${osName}" -e "Linux" ]; then
+    yum install -y openssh-clients
 
-echo "${sshKey}" > key
-chmod 400 key
+    # Get SSH key of cliqruser on node from injected env and stick it in a file.
+    echo "${sshKey}" > key
+    chmod 400 key
 
-tier_ip_varname=CliqrTier_${cliqrAppTierName}_IP
-node_ip=${!tier_ip_varname}
+    # Get the IP Address of the node from variable name that includes the name of the tier.
+    tier_ip_varname=CliqrTier_${cliqrAppTierName}_IP
+    node_ip=${!tier_ip_varname}
 
-mkdir -p ~/.ssh/
-touch ~/.ssh/known_hosts
-ssh-keyscan ${node_ip} >> ~/.ssh/known_hosts
+    # Add the node's ssh key to our list of known hosts to avoid being prompted.
+    mkdir -p ~/.ssh/
+    touch ~/.ssh/known_hosts
+    ssh-keyscan ${node_ip} >> ~/.ssh/known_hosts
 
-curl -o https://raw.githubusercontent.com/datacenter/cloudcenter-content/appd/apps/appd/appd-service-centos.sh
+    # Download the script that you want to run.
+    curl -o script.sh https://raw.githubusercontent.com/datacenter/cloudcenter-content/appd/apps/appd/appd-service-centos.sh
 
-ssh -i key cliqruser@${node_ip} 'bash -s' < appd-service-centos.sh
+    # Run the script.
+    ssh -i key cliqruser@${node_ip} 'bash -s' < script.sh
+
+else
+    # Remote execution isn't supported on Windows right now.
+    print_log "Not Linux, so skipping appd agent install."
+fi
